Updated landing page + Refactored order of pages
I updated the landing page to better represent our project and also re-ordered the pages so that the website will start on the landing page and require a login to access the features of the site. Some various things still need updating such as the layout.tsx in the login folder being auto-generated when refactoring pages.
